Add a utility to merge memory-backed linear counters
What will happen if one machine can not handle all the data that we
need to count. Only solution is distribute the workload to mutiple
machines. And for merging use bit-or operation over individual
bit-sets.

Why this works?
By design algorithm takes care of collisions. It uses unset bits for
couting. Hence bitwise-or will return us equivalent of what we would
have achieved on a single machine.